+++ app/src/main/java/woowacourse/shopping/data/product/ProductRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data.product
 
+import woowacourse.shopping.data.PagedResult
 import woowacourse.shopping.domain.Product
-import woowacourse.shopping.view.PagedResult
 
 interface ProductRepository {
     fun getAll(): List<Product>
